#10 change tab color and icons
