+++ animeapp/constants/index.ts
@@ -1,2 +1,3 @@
 export { default as Theme, Box, Text } from './Theme'
 export { default as icons } from './icons'
+export { default as palette } from './palette'
