+++ animeapp/constants/palette.ts
@@ -1,0 +1,17 @@
+const palette = {
+    lightBlue: '#1CA8E2',
+    mutedBlue: '#159DD5',
+    shyBlue: '#1E3548',
+    darkBlue: '#14242A',
+
+    black: '#000000',
+    white: '#FFFFFF',
+    red: '#FF3B30',
+    green: '#34C759',
+    orange: '#FD823D',
+    yellow: '#FFC61D',
+    gray: '#6D7278'
+    
+}
+
+export default palette
